Incompatible type fix

+++ src/test.c
@@ -19,7 +19,7 @@
 #define DETAILS  false
 
 // Master
-void request_data_sender(uint8_t* data, uint8_t len);
+void request_data_sender(uint8_t* data, uint32_t len);
 void response_packet_handler(modbus_response_t* packet);
 void master_internal_error_handler(void);
 
@@ -324,7 +324,7 @@ void print_test_name(const char* name, uint16_t counter)
 #endif
 }
 
-void request_data_sender(uint8_t* data, uint8_t len)
+void request_data_sender(uint8_t* data, uint32_t len)
 {
 #ifdef __GNUC__
 #if DETAILS
